refactor(mf2): Rename src/runtime/ as src/functions/, add src/formatted-parts.ts
