+++ packages/mf2-messageformat/src/data-model/resolve-expression.ts
@@ -1,10 +1,10 @@
-import type { Context } from '../format-context';
-import type { MessageValue } from '../runtime';
-import type { Expression } from './types';
+import type { Context } from '../format-context.js';
+import type { MessageValue } from '../functions/index.js';
 import { resolveFunctionAnnotation } from './resolve-function-annotation.js';
 import { resolveLiteral } from './resolve-literal.js';
 import { resolveUnsupportedAnnotation } from './resolve-unsupported-annotation.js';
 import { resolveVariableRef } from './resolve-variable.js';
+import type { Expression } from './types.js';
 
 export function resolveExpression(
   ctx: Context,
