+++ packages/mf2-messageformat/src/data-model/resolve-function-annotation.ts
@@ -1,13 +1,14 @@
 import { MessageError } from '../errors.js';
 import type { Context } from '../format-context.js';
-import { MessageFunctionContext, fallback } from '../runtime/index.js';
+import { fallback } from '../functions/fallback.js';
+import { MessageFunctionContext } from '../functions/index.js';
+import { getValueSource, resolveValue } from './resolve-value.js';
 import type {
   FunctionAnnotation,
   Literal,
   Option,
   VariableRef
 } from './types.js';
-import { getValueSource, resolveValue } from './resolve-value.js';
 
 export function resolveFunctionAnnotation(
   ctx: Context,
