+++ packages/mf2-messageformat/src/data-model/resolve-unsupported-annotation.ts
@@ -1,8 +1,8 @@
 import { MessageResolutionError } from '../errors.js';
 import type { Context } from '../format-context.js';
-import { fallback } from '../runtime/index.js';
-import type { Literal, UnsupportedAnnotation, VariableRef } from './types.js';
+import { fallback } from '../functions/fallback.js';
 import { getValueSource } from './resolve-value.js';
+import type { Literal, UnsupportedAnnotation, VariableRef } from './types.js';
 
 export function resolveUnsupportedAnnotation(
   ctx: Context,
