+++ packages/mf2-messageformat/src/functions/datetime.ts
@@ -1,9 +1,6 @@
 import { MessageResolutionError } from '../errors.js';
-import type {
-  MessageExpressionPart,
-  MessageFunctionContext,
-  MessageValue
-} from './index.js';
+import type { MessageFunctionContext, MessageValue } from './index.js';
+import type { MessageExpressionPart } from '../formatted-parts.js';
 import {
   asBoolean,
   asPositiveInteger,
