+++ packages/mf2-messageformat/src/functions/string.ts
@@ -1,8 +1,5 @@
-import type {
-  MessageExpressionPart,
-  MessageFunctionContext,
-  MessageValue
-} from './index.js';
+import type { MessageFunctionContext, MessageValue } from './index.js';
+import type { MessageExpressionPart } from '../formatted-parts.js';
 import { mergeLocales } from './utils.js';
 
 export interface MessageString extends MessageValue {
