+++ packages/mf2-messageformat/src/functions/utils.ts
