+++ packages/mf2-messageformat/src/index.ts
@@ -1,5 +1,8 @@
-export { asDataModel, parseMessage, type CST } from './cst-parser/index.js';
 export type * from './data-model/types.js';
+export type * from './formatted-parts.js';
+export type * from './functions/index.js';
+
+export { asDataModel, parseMessage, type CST } from './cst-parser/index.js';
 export * from './data-model/type-guards.js';
 export { validate } from './data-model/validate.js';
 export { visit } from './data-model/visit.js';
@@ -9,6 +12,12 @@ export {
   MessageResolutionError,
   MessageSyntaxError
 } from './errors.js';
+export { defaultFunctions, utilFunctions } from './functions/index.js';
+export {
+  asBoolean,
+  asPositiveInteger,
+  asString,
+  mergeLocales
+} from './functions/utils.js';
 export { MessageFormat, MessageFormatOptions } from './messageformat.js';
 export { stringifyMessage } from './stringifier/message.js';
-export * from './runtime/index.js';
